[refactor] sidebar 리팩토링
