+++ src/pages/album/[year]/index.tsx
@@ -1,5 +1,5 @@
-import PageTitle from '@src/components/Layout/PageTitle/PageTitle';
-import SideBar from '@src/components/Layout/SideBar/SideBar';
+import PageTitle from '@src/components/Layout/PageTitle';
+import SideBar from '@src/components/Layout/SideBar';
 import CommonCenterWrapper from '@src/components/Layout/CommonCenterWrapper/CommonCenterWrapper';
 
 export default function Year() {
